+++ examples/with-miniapp-plugin-component/project.config.json
@@ -2,7 +2,33 @@
 	"miniprogramRoot": "build/wechat-miniprogram/",
 	"pluginRoot": "wechat-miniprogram-plugin/",
 	"compileType": "plugin",
-	"setting": {},
+	"setting": {
+		"urlCheck": true,
+		"es6": true,
+		"enhance": false,
+		"postcss": true,
+		"preloadBackgroundData": false,
+		"minified": true,
+		"newFeature": false,
+		"coverView": true,
+		"nodeModules": false,
+		"autoAudits": false,
+		"showShadowRootInWxmlPanel": true,
+		"scopeDataCheck": false,
+		"uglifyFileName": false,
+		"checkInvalidKey": true,
+		"checkSiteMap": true,
+		"uploadWithSourceMap": true,
+		"compileHotReLoad": false,
+		"babelSetting": {
+			"ignore": [],
+			"disablePlugins": [],
+			"outputPath": ""
+		},
+		"useIsolateContext": true,
+		"useCompilerModule": true,
+		"userConfirmedUseCompilerModuleSwitch": false
+	},
 	"appid": "wxc0575284fd6f4091",
 	"projectname": "with-miniapp-plugin-component",
 	"simulatorType": "wechat",
