--- packages/miniapp-render/src/utils/checkEventAccessDomNode.js
@@ -1,13 +1,0 @@
-export default function(evt, domNode, dest) {
-  const targetNode = dest || domNode.ownerDocument.body;
-  let target = evt.target;
-
-  if (domNode === targetNode) return true;
-
-  while (target && target !== dest) {
-    if (target === domNode) return true;
-    target = target.parentNode;
-  }
-
-  return false;
-}
